requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autobahn==25.11.1
Automat==25.4.16
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
cbor2==5.7.1
certifi==2024.8.30
cffi==2.0.0
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
daphne==4.2.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1
django-auditlog==3.4.0
django-auth-adfs==1.14.0
django-simple-history==3.10.1
django-sslserver==0.22
djangorestframework==3.15.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
firebase-admin==6.7.0
fonttools==4.55.0
fqdn==1.5.1
getmac==0.9.5
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==25.3.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pyudev==0.24.3
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
rsa==4.9
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.15.0
tzdata==2024.1
ujson==5.11.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.13
WMI==1.5.1
zope.interface==8.1.1
